created ads views, serializers. Do permissions

+++ skymarket/ads/models.py
@@ -10,7 +10,7 @@ from users.models import User
 class Ad(models.Model):
     image = models.ImageField(upload_to="image/", null=True)
     title = models.CharField(max_length=200)
-    price = models.PositiveSmallIntegerField()
+    price = models.PositiveIntegerField()
     description = models.CharField(max_length=1000)
     author = models.ForeignKey(User, on_delete=models.CASCADE)
     created_at = models.DateTimeField(auto_now_add=True, blank=True)
